requirements.txt
flask==3.1.0
twilio==9.4.0
python-dotenv==1.1.0
gunicorn==23.0.0
